separate TopN and Ordered StateCache impls
